% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_plot_ly.R
\name{easy_plot_ly}
\alias{easy_plot_ly}
\title{A shortcut for generating plots with plot_ly}
\usage{
easy_plot_ly(
  x = NULL,
  y = NULL,
  z = NULL,
  color = NULL,
  type = ...,
  data = NULL
)
}
\arguments{
\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{z}{The z-axis variable}

\item{color}{A cateogrical or quantitative varaible to subset the data}

\item{type}{The type of plot you would like to generate: scatter, line,
surface - only takes a z-axis matrix, 3d_density, mesh, or specify nothing to
autogenerate a plot}

\item{data}{The datset that will be used}
}
\value{
A plot
}
\description{
A shortcut for generating plots with plot_ly
}
