% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_scatter.R
\name{ez_scatter}
\alias{ez_scatter}
\title{Shortcut functions for generating scatterplots from ggplot2}
\usage{
ez_scatter(data, x_num, y_num, z_cat = NULL, color = NULL, shape = NULL)
}
\arguments{
\item{x_num}{a numeric variable string for the x-axis}

\item{y_num}{a numeric variable string for the y-axis}

\item{z_cat}{an optional categorical variable sting for different groups}

\item{color}{color string for graph. If the scatterplot has two variables the color must be chosen from ggplot2 colors. If the column is three variables the color must be chosen from RcolorBrewer.}

\item{shape}{if the scatterplot has two variables, input a number corresponding to a point shape in R. If the scatterplot has three variablles, input the catigorical variable string.}
}
\value{
A plot
}
\description{
Shortcut functions for generating scatterplots from ggplot2
}
