% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_choropleth.R
\name{leafletPlot}
\alias{leafletPlot}
\title{A shortcut for generating choropleth maps}
\usage{
leafletPlot(
  data_set,
  col_name,
  col_variable,
  type_of_map,
  bins_of_map,
  col_variable_unit = "",
  title = "Chloropleth Map"
)
}
\arguments{
\item{data_set}{The name of the data set}

\item{col_name}{The column name in the data set containing the regions (must be in quotes)}

\item{col_variable}{The column name containing the variable of interest (must be in quotes)}

\item{type_of_map}{The type of map you wish to specify; possible options are "countries", "us_states"}

\item{bins_of_map}{The type of bins user wishes to use for map; must be a vector}

\item{col_variable_unit}{The units of your variable; must be in quotes; if it is a percentage, put this argument as "\%"}

\item{title}{The title of the chloropleth map; must be in quotes}
}
\value{
Leaflet Map
}
\description{
A shortcut for generating choropleth maps
}
