% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_plot_ly.R
\name{theme_ly}
\alias{theme_ly}
\title{Edits plot aesthetics}
\usage{
theme_ly(
  x_axis = NULL,
  y_axis = NULL,
  z_axis = NULL,
  title = NULL,
  colorscale_title = NULL,
  legend = FALSE,
  plot
)
}
\arguments{
\item{x_axis}{The x-axis label}

\item{y_axis}{The y-axis label}

\item{z_axis}{The z-axis label}

\item{title}{The plot title}

\item{colorscale_title}{The title for the colorscale}

\item{legend}{Show legend}

\item{plot}{The plot that you want to add a plane to}
}
\value{
aesthetics adjustments for a plot
}
\description{
Edits plot aesthetics
}
