% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_plot_ly.R
\name{add_reg_plane}
\alias{add_reg_plane}
\title{3D Regression Plane}
\usage{
Recommended that you use this function with a scatter plot

add_reg_plane(x, y, z , plot, color = 'blue')
}
\arguments{
\item{x}{The x-axis variable}

\item{y}{The y-axis variable}

\item{z}{The z-axis variable}

\item{plot}{The plot that you want to add a plane to}

\item{color}{The color of the plane}
}
\value{
A 3D plane added to an exisiting plot
}
\description{
This is a helper function to add a 3D regression plane to a plot
}
\examples{
newPlot <- easy_plot_ly(x = iris$Sepal.Length, y = iris$Sepal.Width, z = iris$Petal.Length,
             color = iris$Species, type = 'scatter', data = iris)
add_reg_plane(x = iris$Sepal.Length, y = iris$Sepal.Width, z = iris$Petal.Length, plot = newPlot)


}
