% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_func.R
\name{ez_scatter}
\alias{ez_scatter}
\title{Shortcut functions for generating scatterplots from ggplot2}
\usage{
ez_scatter(data, x_num, y_num, z_cat = NULL, color = NULL)
}
\arguments{
\item{x_num}{numeric variable for x-axis}

\item{y_num}{numeric variable for y-axis}

\item{z_cat}{optional categorical variable for different groups}

\item{color}{color of the z_cat variable
Defaults to \code{TRUE}.}
}
\value{
A plot
}
\description{
Shortcut functions for generating scatterplots from ggplot2
}
