% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_plots.R
\name{mult_plots}
\alias{mult_plots}
\title{A shortcut for generating multiple plots side by side using the package patchwork}
\usage{
mult_plots(
  p1 = NULL,
  p2 = NULL,
  p3 = NULL,
  p4 = NULL
)
}
\arguments{
\item{p1}{The first plot}

\item{p2}{The second plot}

\item{p3}{The third plot}

\item{p4}{The fourth plot}

mult_plots(p1,p2,p3,p4)}
}
\value{
Multiple plots side by side
}
\description{
A shortcut for generating multiple plots side by side using the package patchwork
}
